+++ player/amiga/mrplay.c
@@ -2251,7 +2251,7 @@ int main(int argc, char **argv)
         printf("usage: mrplay [--user-agent <value>] [--referer <value>] "
                "<file.avi|file.mov|file.ts|file.m2ts|"
                "file.mjpeg|file.m4v> "
-               "[--aga] [--aga-window] [--ham] [--ham6] [--p96] "
+               "[--aga] [--aga-window] [--aga-window-half] [--ham] [--ham6] [--p96] "
                "[--2x] [--lace] [--ecs-fast] [--ecs32] [--ehb] [--copper-vdouble] "
                "[--loop] "
                "[--wpa|--c2p|--riva-c2p|--kalms-c2p|--direct-c2p] "
@@ -2285,6 +2285,7 @@ int main(int argc, char **argv)
             else if (!strcmp(argv[i], "--p96"))  display_set_force_p96(1);
             else if (!strcmp(argv[i], "--rtg-half")) rtg_half = 1;
             else if (!strcmp(argv[i], "--aga-window")) display_set_aga_window(1);
+            else if (!strcmp(argv[i], "--aga-window-half")) display_set_aga_window(2);
             else if (!strcmp(argv[i], "--ham"))  display_set_ham(8);
             else if (!strcmp(argv[i], "--ham6")) display_set_ham(6);
             else if (!strcmp(argv[i], "--2x"))   display_set_scale(2);